Documentation for multi-object codec service. Add updated changes.md for 0.4.2
