--- sdk/python/CHANGES.txt
@@ -1,19 +1,0 @@
-**2016-04-15 version 0.4.0:**
-
-  * Introduced netconf service and codec service
-    * Netconf service provides APIs to execute netconf operations
-    * Codec service provides APIs to encode python objects and decode payloads
-  * Support for yang deviation
-  * Support for subscribing to model-driven telemetry
-  * Logging made more consistent
-    * CRUDService outputs type of operation
-    * When logging is enabled, all NETCONF messages are logged including commit
-    * Log messages at various stages (send RPC request, receive reply, commit 
-      etc) instead of logging all at once at the end
-  * Updated enums in YDK classes to use enum34
-    * Improved enum documentation
-  * Improved error reporting for ydk-py and ydk-gen
-
-**2016-03-11 version 0.3.0:**
-
-  * First public release.
